refactor(telegram): extract media sends into TelegramMediaMixin (adapter god-file slice A4)
(cherry picked from commit d9e87873ec0e19e3379a28e414d68eedf91cfbbe)
