+++ plugins/platforms/telegram/adapter.py
@@ -285,6 +285,12 @@ from gateway.platforms.base import (
 from plugins.platforms.telegram.telegram_ids import (
     normalize_telegram_chat_id,
 )
+from plugins.platforms.telegram.telegram_media import (
+    TelegramMediaMixin,
+    _MEDIA_SEND_READ_TIMEOUT,
+    _coerce_duration_seconds,
+    _probe_voice_duration_seconds,
+)
 from plugins.platforms.telegram.telegram_network import (
     TelegramFallbackTransport,
     discover_fallback_ips,
@@ -306,74 +312,6 @@ from plugins.platforms.telegram.telegram_inbound import (
     _redact_telegram_error_text,
 )
 from plugins.platforms.telegram.telegram_reactions import TelegramReactionsMixin
-
-def _coerce_duration_seconds(value: Any) -> Optional[int]:
-    """Round a raw length to whole positive seconds, or None if unusable."""
-    try:
-        secs = int(round(float(value)))
-    except (TypeError, ValueError):
-        return None
-    return secs if secs > 0 else None
-
-
-def _probe_voice_duration_seconds(path: str) -> Optional[int]:
-    """Best-effort audio length in whole seconds for outgoing voice/audio.
-
-    Telegram only auto-derives a clip's duration from container metadata for
-    short recordings; longer ones (roughly 5 min+) are sent with duration 0
-    and render as ``0:00`` in the player. We read the length locally and pass
-    it explicitly so the bubble shows the real time.
-
-    Mirrors ``gateway.run._probe_audio_duration``: stdlib ``wave`` for WAV,
-    then mutagen for OGG/Opus/MP3/M4A metadata, then an ``ffprobe`` fallback.
-    All three are optional — when none can read the file we return ``None``
-    and the caller omits ``duration``, falling back to Telegram's own
-    (possibly absent) metadata, i.e. the prior behavior. Blocking (mutagen
-    read + ffprobe subprocess), so call it via ``asyncio.to_thread``.
-    """
-    ext = os.path.splitext(path)[1].lower()
-
-    if ext == ".wav":
-        try:
-            import wave
-
-            with wave.open(path, "rb") as wf:
-                rate = wf.getframerate() or 0
-                if rate:
-                    secs = _coerce_duration_seconds(wf.getnframes() / float(rate))
-                    if secs is not None:
-                        return secs
-        except Exception:
-            pass
-
-    try:
-        import mutagen
-
-        audio = mutagen.File(path)
-        secs = _coerce_duration_seconds(
-            getattr(getattr(audio, "info", None), "length", None)
-        )
-        if secs is not None:
-            return secs
-    except Exception:
-        pass
-
-    try:
-        import shutil
-        import subprocess
-
-        if shutil.which("ffprobe"):
-            proc = subprocess.run(
-                ["ffprobe", "-v", "error", "-show_entries", "format=duration",
-                 "-of", "default=noprint_wrappers=1:nokey=1", path],
-                capture_output=True, text=True, encoding="utf-8", errors="replace", timeout=5,
-            )
-            if proc.returncode == 0:
-                return _coerce_duration_seconds(proc.stdout.strip())
-    except Exception:
-        pass
-
-    return None
 
 
 def check_telegram_requirements() -> bool:
@@ -532,13 +470,6 @@ _POLLING_ERROR_TASK_STUCK_TIMEOUT = 300.0
 # A generation is not healthy until the dedicated getUpdates request returns
 # successfully. This exceeds a normal long-poll cycle for healthy idle bots.
 _POLLING_PROGRESS_TIMEOUT = 60.0
-# Telegram transcodes an uploaded video before it answers sendVideo, so the
-# wait for the response is unrelated to how fast the bytes went out and can
-# outlast the 20s read timeout the rest of the Bot API is tuned for. Only
-# media sends take this longer budget; ordinary calls keep the short one so a
-# dead request is still noticed quickly. Kept modest deliberately — this is
-# also how long a user waits to be told the attachment failed.
-_MEDIA_SEND_READ_TIMEOUT = 60.0
 _POLLING_GENERATION_CONTEXT: ContextVar[Optional[int]] = ContextVar(
     "telegram_polling_generation", default=None
 )
@@ -548,7 +479,7 @@ class _PollingLifecycleAbort(RuntimeError):
     """Internal control flow for polling startup fenced by teardown."""
 
 
-class TelegramAdapter(TelegramLifecycleMixin, TelegramReactionsMixin, TelegramPollingMixin, TelegramIngestMixin, TelegramTextDeliveryMixin, TelegramRichMixin, BasePlatformAdapter):
+class TelegramAdapter(TelegramMediaMixin, TelegramLifecycleMixin, TelegramReactionsMixin, TelegramPollingMixin, TelegramIngestMixin, TelegramTextDeliveryMixin, TelegramRichMixin, BasePlatformAdapter):
     """
     Telegram bot adapter.
 
@@ -3122,802 +3053,6 @@ class TelegramAdapter(TelegramLifecycleMixin, TelegramReactionsMixin, TelegramPo
                 await query.edit_message_text(text=appended, reply_markup=None)
         except Exception:
             pass
-
-    def _missing_media_path_error(self, label: str, path: str) -> str:
-        """Build an actionable file-not-found error for gateway MEDIA delivery.
-
-        Paths like /workspace/... or /output/... often only exist inside the
-        Docker sandbox, while the gateway process runs on the host.
-        """
-        error = f"{label} file not found: {path}"
-        if path.startswith(("/workspace/", "/output/", "/outputs/")):
-            error += (
-                " (path may only exist inside the Docker sandbox. "
-                "Bind-mount a host directory and emit the host-visible "
-                "path in MEDIA: for gateway file delivery.)"
-            )
-        return error
-
-    def _telegram_media_too_large_note(self, label: str, file_size: Any, max_bytes: int) -> str:
-        limit_mb = max(1, max_bytes // (1024 * 1024))
-        try:
-            size_mb = int(file_size or 0) / (1024 * 1024)
-            size_text = f"{size_mb:.1f} MB"
-        except (TypeError, ValueError):
-            size_text = "unknown size"
-        return (
-            f"[Telegram {label} skipped: file size {size_text} exceeds the "
-            f"{limit_mb} MB limit. Ask the user to send a smaller file.]"
-        )
-
-    def _telegram_media_size_allowed(self, source: Any, label: str) -> tuple[bool, Optional[str]]:
-        """Validate Telegram media size before downloading into memory."""
-        max_bytes = int(getattr(self, "_max_doc_bytes", 20 * 1024 * 1024) or 20 * 1024 * 1024)
-        file_size = getattr(source, "file_size", None)
-        try:
-            size = int(file_size or 0)
-        except (TypeError, ValueError):
-            size = 0
-        if size <= 0:
-            return True, None
-        if size <= max_bytes:
-            return True, None
-        return False, self._telegram_media_too_large_note(label, size, max_bytes)
-
-    async def send_voice(
-        self,
-        chat_id: str,
-        audio_path: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send audio as a native Telegram voice message or audio file."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-        
-        try:
-            if not os.path.exists(audio_path):
-                return SendResult(success=False, error=self._missing_media_path_error("Audio", audio_path))
-            
-            # Compute duration locally — Telegram drops it for long clips
-            # (~5 min+), which then show 0:00 in the player.
-            _duration_secs = await asyncio.to_thread(
-                _probe_voice_duration_seconds, audio_path
-            )
-
-            # Render caption markdown (#32029): auto-TTS captions carry the
-            # agent's markdown reply, which showed literal *asterisks* and
-            # [links](...) without a parse_mode. Format to MarkdownV2 when it
-            # fits the 1024-char caption cap; fall back to the raw text
-            # (previous behaviour) when formatting would overflow or the
-            # Bot API rejects the entities.
-            _caption_variants: List[tuple] = []
-            if caption:
-                try:
-                    _formatted_caption = self.format_message(caption)
-                    if utf16_len(_formatted_caption) <= 1024:
-                        _caption_variants.append(
-                            (_formatted_caption, ParseMode.MARKDOWN_V2)
-                        )
-                except Exception:
-                    logger.debug(
-                        "[%s] voice caption MarkdownV2 formatting failed; "
-                        "sending plain caption", self.name, exc_info=True,
-                    )
-                _caption_variants.append((caption[:1024], None))
-            else:
-                _caption_variants.append((None, None))
-
-            with open(audio_path, "rb") as audio_file:
-                ext = os.path.splitext(audio_path)[1].lower()
-                # .ogg / .opus files -> send as voice (round playable bubble)
-                if ext in {".ogg", ".opus"}:
-                    _voice_thread = self._metadata_thread_id(metadata)
-                    reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-                    voice_thread_kwargs = self._thread_kwargs_for_send(
-                        chat_id,
-                        _voice_thread,
-                        metadata,
-                        reply_to_message_id=reply_to_id,
-                        reply_to_mode=self._reply_to_mode
-                    )
-                    msg = None
-                    _last_parse_error: Optional[Exception] = None
-                    for _cap_text, _cap_parse_mode in _caption_variants:
-                        try:
-                            msg = await self._send_with_dm_topic_reply_anchor_retry(
-                                self._bot.send_voice,
-                                {
-                                    "chat_id": normalize_telegram_chat_id(chat_id),
-                                    "voice": audio_file,
-                                    "caption": _cap_text,
-                                    "parse_mode": _cap_parse_mode,
-                                    "reply_to_message_id": reply_to_id,
-                                    "duration": _duration_secs,
-                                    "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                                    **voice_thread_kwargs,
-                                    **self._notification_kwargs(metadata),
-                                },
-                                metadata,
-                                reply_to_id,
-                                "voice",
-                                reset_media=lambda: audio_file.seek(0),
-                            )
-                            break
-                        except Exception as _cap_error:
-                            # Only retry the next (plain) variant on entity
-                            # parse failures; anything else is a real send
-                            # error for the outer handler.
-                            if (_cap_parse_mode is not None
-                                    and ("parse" in str(_cap_error).lower()
-                                         or "entit" in str(_cap_error).lower())):
-                                logger.warning(
-                                    "[%s] voice caption MarkdownV2 rejected, "
-                                    "retrying plain: %s",
-                                    self.name,
-                                    _redact_telegram_error_text(_cap_error),
-                                )
-                                _last_parse_error = _cap_error
-                                audio_file.seek(0)
-                                continue
-                            raise
-                    if msg is None:
-                        raise _last_parse_error or RuntimeError(
-                            "Telegram send_voice failed for all caption variants"
-                        )
-                elif ext in {".mp3", ".m4a"}:
-                    # Telegram's Bot API sendAudio only accepts MP3 / M4A.
-                    _audio_thread = self._metadata_thread_id(metadata)
-                    reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-                    audio_thread_kwargs = self._thread_kwargs_for_send(
-                        chat_id,
-                        _audio_thread,
-                        metadata,
-                        reply_to_message_id=reply_to_id,
-                        reply_to_mode=self._reply_to_mode
-                    )
-                    msg = await self._send_with_dm_topic_reply_anchor_retry(
-                        self._bot.send_audio,
-                        {
-                            "chat_id": normalize_telegram_chat_id(chat_id),
-                            "audio": audio_file,
-                            "caption": caption[:1024] if caption else None,
-                            "reply_to_message_id": reply_to_id,
-                            "duration": _duration_secs,
-                            "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                            **audio_thread_kwargs,
-                            **self._notification_kwargs(metadata),
-                        },
-                        metadata,
-                        reply_to_id,
-                        "audio",
-                        reset_media=lambda: audio_file.seek(0),
-                    )
-                else:
-                    # Formats Telegram can't play natively (.wav, .flac, ...)
-                    # — fall back to document delivery instead of raising.
-                    return await self.send_document(
-                        chat_id=chat_id,
-                        file_path=audio_path,
-                        caption=caption,
-                        reply_to=reply_to,
-                        metadata=metadata,
-                    )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.error(
-                "[%s] Failed to send Telegram voice/audio, falling back to base adapter: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            return await super().send_voice(chat_id, audio_path, caption, reply_to, metadata=metadata)
-
-    async def send_multiple_images(
-        self,
-        chat_id: str,
-        images: List[tuple],
-        metadata: Optional[Dict[str, Any]] = None,
-        human_delay: float = 0.0,
-    ) -> None:
-        """Send a batch of images natively via Telegram's media group API.
-
-        Telegram's ``send_media_group`` bundles up to 10 photos/videos into
-        a single album. Larger batches are chunked. Animated GIFs cannot
-        go into a media group (they require ``send_animation``), so they
-        are peeled off and sent individually via the base default path.
-
-        URL-based photos go into the group directly; local files are
-        opened as byte streams. On failure the whole batch falls back to
-        the base adapter's per-image loop.
-        """
-        if not self._bot:
-            return
-        if not images:
-            return
-
-        try:
-            from telegram import InputMediaPhoto
-        except Exception as exc:  # pragma: no cover - missing SDK
-            logger.warning(
-                "[%s] InputMediaPhoto unavailable, falling back to per-image send: %s",
-                self.name, exc,
-            )
-            await super().send_multiple_images(chat_id, images, metadata, human_delay)
-            return
-
-        # Peel off animations — they need send_animation, not send_media_group
-        animations: List[tuple] = []
-        photos: List[tuple] = []
-        for image_url, alt_text in images:
-            if not image_url.startswith("file://") and self._is_animation_url(image_url):
-                animations.append((image_url, alt_text))
-            else:
-                photos.append((image_url, alt_text))
-
-        # Animations: route through the base default (per-image send_animation)
-        if animations:
-            await super().send_multiple_images(
-                chat_id, animations, metadata, human_delay=human_delay,
-            )
-
-        if not photos:
-            return
-
-        from urllib.parse import unquote as _unquote
-        _thread = self._metadata_thread_id(metadata)
-
-        # Chunk into groups of 10 (Telegram's album limit)
-        CHUNK = 10
-        chunks = [photos[i:i + CHUNK] for i in range(0, len(photos), CHUNK)]
-
-        for chunk_idx, chunk in enumerate(chunks):
-            if human_delay > 0 and chunk_idx > 0:
-                await asyncio.sleep(human_delay)
-
-            media: List[Any] = []
-            opened_files: List[Any] = []
-            try:
-                for image_url, alt_text in chunk:
-                    caption = alt_text[:1024] if alt_text else None
-                    if image_url.startswith("file://"):
-                        local_path = _unquote(image_url[7:])
-                        if not os.path.exists(local_path):
-                            logger.warning(
-                                "[%s] Skipping missing image in media group: %s",
-                                self.name, local_path,
-                            )
-                            continue
-                        fh = open(local_path, "rb")
-                        opened_files.append(fh)
-                        media.append(InputMediaPhoto(media=fh, caption=caption))
-                    else:
-                        media.append(InputMediaPhoto(media=image_url, caption=caption))
-
-                if not media:
-                    continue
-
-                logger.info(
-                    "[%s] Sending media group of %d photo(s) (chunk %d/%d)",
-                    self.name, len(media), chunk_idx + 1, len(chunks),
-                )
-                reply_to_id = self._reply_to_message_id_for_send(None, metadata, reply_to_mode=self._reply_to_mode)
-                thread_kwargs = self._thread_kwargs_for_send(
-                    chat_id,
-                    _thread,
-                    metadata,
-                    reply_to_message_id=reply_to_id,
-                    reply_to_mode=self._reply_to_mode
-                )
-
-                def _reset_opened_files() -> None:
-                    for fh in opened_files:
-                        try:
-                            fh.seek(0)
-                        except Exception:
-                            pass
-
-                await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_media_group,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "media": media,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "media group",
-                    reset_media=_reset_opened_files,
-                )
-            except Exception as e:
-                logger.warning(
-                    "[%s] send_media_group failed (chunk %d/%d), falling back to per-image: %s",
-                    self.name, chunk_idx + 1, len(chunks), _redact_telegram_error_text(e),
-                    exc_info=True,
-                )
-                # Fallback: send each photo in this chunk individually
-                await super().send_multiple_images(
-                    chat_id, chunk, metadata, human_delay=human_delay,
-                )
-            finally:
-                for fh in opened_files:
-                    try:
-                        fh.close()
-                    except Exception:
-                        pass
-
-    async def send_image_file(
-        self,
-        chat_id: str,
-        image_path: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send a local image file natively as a Telegram photo."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        try:
-            if not os.path.exists(image_path):
-                return SendResult(success=False, error=self._missing_media_path_error("Image", image_path))
-
-            _thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            with open(image_path, "rb") as image_file:
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_photo,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "photo": image_file,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "photo",
-                    reset_media=lambda: image_file.seek(0),
-                )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            error_str = str(e)
-            # Dimension-related errors are the expected case for valid image
-            # files that Telegram just refuses as photos (screenshots, extreme
-            # aspect ratios). Log at INFO because the document fallback is
-            # the correct path. Any other send_photo failure also falls back
-            # to document (rate limits, corrupt file markers, format edge
-            # cases), but at WARNING because it's unexpected and worth
-            # surfacing in logs.
-            is_dim_error = (
-                "Photo_invalid_dimensions" in error_str
-                or "PHOTO_INVALID_DIMENSIONS" in error_str
-            )
-            if is_dim_error:
-                logger.info(
-                    "[%s] Image dimensions exceed Telegram photo limits, "
-                    "sending as document: %s",
-                    self.name,
-                    image_path,
-                )
-            else:
-                logger.warning(
-                    "[%s] Failed to send Telegram local image as photo, "
-                    "trying document fallback: %s",
-                    self.name,
-                    _redact_telegram_error_text(e),
-                    exc_info=True,
-                )
-            # Fallback to sending as document (file) — no dimension limit,
-            # only 50MB size limit. If even that fails, fall back to the
-            # base adapter's text-only "Image: /path" rendering.
-            try:
-                return await self.send_document(
-                    chat_id=chat_id,
-                    file_path=image_path,
-                    caption=caption,
-                    file_name=os.path.basename(image_path),
-                    reply_to=reply_to,
-                    metadata=metadata,
-                )
-            except Exception as doc_err:
-                logger.error(
-                    "[%s] Failed to send Telegram local image as document, "
-                    "falling back to base adapter: %s",
-                    self.name,
-                    doc_err,
-                    exc_info=True,
-                )
-                return await super().send_image_file(chat_id, image_path, caption, reply_to, metadata=metadata)
-
-    async def send_document(
-        self,
-        chat_id: str,
-        file_path: str,
-        caption: Optional[str] = None,
-        file_name: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send a document/file natively as a Telegram file attachment."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        try:
-            if not os.path.exists(file_path):
-                return SendResult(success=False, error=self._missing_media_path_error("File", file_path))
-
-            display_name = file_name or os.path.basename(file_path)
-            _thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-
-            with open(file_path, "rb") as f:
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_document,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "document": f,
-                        "filename": display_name,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "document",
-                    reset_media=lambda: f.seek(0),
-                )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.warning(
-                "[%s] Failed to send document: %s",
-                self.name, _redact_telegram_error_text(e),
-            )
-            return await super().send_document(chat_id, file_path, caption, file_name, reply_to, metadata=metadata)
-
-    async def send_video(
-        self,
-        chat_id: str,
-        video_path: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send a video natively as a Telegram video message."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        try:
-            if not os.path.exists(video_path):
-                return SendResult(success=False, error=self._missing_media_path_error("Video", video_path))
-
-            _thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            with open(video_path, "rb") as f:
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_video,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "video": f,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "video",
-                    reset_media=lambda: f.seek(0),
-                )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.warning(
-                "[%s] Failed to send video: %s",
-                self.name, _redact_telegram_error_text(e),
-            )
-            return await super().send_video(chat_id, video_path, caption, reply_to, metadata=metadata)
-
-    async def send_image(
-        self,
-        chat_id: str,
-        image_url: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-    ) -> SendResult:
-        """Send an image natively as a Telegram photo.
-        
-        Tries URL-based send first (fast, works for <5MB images).
-        Falls back to downloading and uploading as file (supports up to 10MB).
-        """
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        from tools.url_safety import is_safe_url
-        if not is_safe_url(image_url):
-            logger.warning("[%s] Blocked unsafe image URL (SSRF protection)", self.name)
-            return await super().send_image(chat_id, image_url, caption, reply_to, metadata=metadata)
-
-        try:
-            # Telegram can send photos directly from URLs (up to ~5MB)
-            _photo_thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            photo_thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _photo_thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            msg = await self._send_with_dm_topic_reply_anchor_retry(
-                self._bot.send_photo,
-                {
-                    "chat_id": normalize_telegram_chat_id(chat_id),
-                    "photo": image_url,
-                    "caption": caption[:1024] if caption else None,
-                    "reply_to_message_id": reply_to_id,
-                    "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                    **photo_thread_kwargs,
-                    **self._notification_kwargs(metadata),
-                },
-                metadata,
-                reply_to_id,
-                "URL photo",
-            )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.warning(
-                "[%s] URL-based send_photo failed, trying file upload: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            # Fallback: download and upload as file (supports up to 10MB)
-            try:
-                from gateway.platforms.base import _ssrf_redirect_guard
-                from tools.url_safety import create_ssrf_safe_async_client
-
-                async with create_ssrf_safe_async_client(
-                    timeout=30.0,
-                    event_hooks={"response": [_ssrf_redirect_guard]},
-                ) as client:
-                    resp = await client.get(image_url)
-                    resp.raise_for_status()
-                    image_data = resp.content
-
-                upload_thread_kwargs = self._thread_kwargs_for_send(
-                    chat_id,
-                    _photo_thread,
-                    metadata,
-                    reply_to_message_id=reply_to_id,
-                    reply_to_mode=self._reply_to_mode
-                )
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_photo,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "photo": image_data,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **upload_thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "uploaded photo",
-                )
-                return SendResult(success=True, message_id=str(msg.message_id))
-            except Exception as e2:
-                logger.error(
-                    "[%s] File upload send_photo also failed: %s",
-                    self.name,
-                    e2,
-                    exc_info=True,
-                )
-                # Final fallback: send URL as text
-                return await super().send_image(chat_id, image_url, caption, reply_to, metadata=metadata)
-
-    async def send_animation(
-        self,
-        chat_id: str,
-        animation_url: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-    ) -> SendResult:
-        """Send an animated GIF natively as a Telegram animation (auto-plays inline)."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-        
-        try:
-            _anim_thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            animation_thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _anim_thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            msg = await self._send_with_dm_topic_reply_anchor_retry(
-                self._bot.send_animation,
-                {
-                    "chat_id": normalize_telegram_chat_id(chat_id),
-                    "animation": animation_url,
-                    "caption": caption[:1024] if caption else None,
-                    "reply_to_message_id": reply_to_id,
-                    "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                    **animation_thread_kwargs,
-                    **self._notification_kwargs(metadata),
-                },
-                metadata,
-                reply_to_id,
-                "animation",
-            )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.error(
-                "[%s] Failed to send Telegram animation, falling back to photo: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            # Fallback: try as a regular photo
-            return await self.send_image(chat_id, animation_url, caption, reply_to, metadata=metadata)
-
-    @staticmethod
-    def _is_transient_typing_error(exc: Exception) -> bool:
-        """Return True for Telegram typing errors worth cooling down."""
-        retry_after = getattr(exc, "retry_after", None)
-        if retry_after is not None:
-            return True
-
-        status_code = getattr(exc, "status_code", None) or getattr(exc, "code", None)
-        if isinstance(status_code, int) and (status_code == 429 or status_code >= 500):
-            return True
-
-        text = str(exc).lower()
-        if any(marker in text for marker in ("too many requests", "rate limit", "timed out", "timeout", "temporar")):
-            return True
-        if isinstance(exc, (OSError, TimeoutError, ConnectionError, asyncio.TimeoutError)):
-            return True
-        return False
-
-    def _record_typing_cooldown(self, chat_id: str, exc: Exception) -> None:
-        """Suppress Telegram typing refreshes for this chat after transient failures."""
-        if not hasattr(self, "_telegram_typing_cooldown_until"):
-            self._telegram_typing_cooldown_until = {}
-        loop = asyncio.get_running_loop()
-        retry_after = getattr(exc, "retry_after", None)
-        try:
-            delay = float(retry_after) if retry_after is not None else self._telegram_typing_cooldown_seconds
-        except (TypeError, ValueError):
-            delay = self._telegram_typing_cooldown_seconds
-        delay = max(1.0, min(delay, 300.0))
-        self._telegram_typing_cooldown_until[str(chat_id)] = loop.time() + delay
-
-    def _typing_in_cooldown(self, chat_id: str) -> bool:
-        if not hasattr(self, "_telegram_typing_cooldown_until"):
-            self._telegram_typing_cooldown_until = {}
-            self._telegram_typing_cooldown_seconds = 30.0
-        until = self._telegram_typing_cooldown_until.get(str(chat_id))
-        if until is None:
-            return False
-        if asyncio.get_running_loop().time() < until:
-            return True
-        self._telegram_typing_cooldown_until.pop(str(chat_id), None)
-        return False
-
-    async def send_typing(self, chat_id: str, metadata: Optional[Dict[str, Any]] = None) -> None:
-        """Send typing indicator."""
-        if not self._bot or self._typing_in_cooldown(chat_id):
-            return
-
-        _is_dm_topic: bool = False
-        message_thread_id: Optional[int] = None
-        try:
-            _typing_thread = self._metadata_thread_id(metadata)
-            _is_dm_topic = bool(metadata and metadata.get("telegram_dm_topic_reply_fallback"))
-            message_thread_id = self._message_thread_id_for_typing(_typing_thread)
-            await self._bot.send_chat_action(
-                chat_id=normalize_telegram_chat_id(chat_id),
-                action="typing",
-                message_thread_id=message_thread_id,
-            )
-            self._telegram_typing_cooldown_until.pop(str(chat_id), None)
-        except Exception as e:
-            # For DM topic lanes, Telegram may reject message_thread_id.
-            # Fall back to sending typing without thread_id so the typing
-            # indicator at least appears in the main DM view.
-            if _is_dm_topic and message_thread_id is not None:
-                try:
-                    await self._bot.send_chat_action(
-                        chat_id=normalize_telegram_chat_id(chat_id),
-                        action="typing",
-                    )
-                    self._telegram_typing_cooldown_until.pop(str(chat_id), None)
-                    return
-                except Exception as fallback_exc:
-                    if self._is_transient_typing_error(fallback_exc):
-                        self._record_typing_cooldown(chat_id, fallback_exc)
-            elif self._is_transient_typing_error(e):
-                self._record_typing_cooldown(chat_id, e)
-            # Typing failures are non-fatal; log at debug level only.
-            logger.debug(
-                "[%s] Failed to send Telegram typing indicator: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-
-    async def get_chat_info(self, chat_id: str) -> Dict[str, Any]:
-        """Get information about a Telegram chat."""
-        if not self._bot:
-            return {"name": "Unknown", "type": "dm"}
-        
-        try:
-            chat = await self._bot.get_chat(normalize_telegram_chat_id(chat_id))
-            
-            chat_type = "dm"
-            if chat.type == ChatType.GROUP:
-                chat_type = "group"
-            elif chat.type == ChatType.SUPERGROUP:
-                chat_type = "group"
-                if chat.is_forum:
-                    chat_type = "forum"
-            elif chat.type == ChatType.CHANNEL:
-                chat_type = "channel"
-            
-            return {
-                "name": chat.title or chat.full_name or str(chat_id),
-                "type": chat_type,
-                "username": chat.username,
-                "is_forum": getattr(chat, "is_forum", False),
-            }
-        except Exception as e:
-            logger.error(
-                "[%s] Failed to get Telegram chat info for %s: %s",
-                self.name,
-                chat_id,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            return {"name": str(chat_id), "type": "dm", "error": str(e)}
 
     def format_message(self, content: str) -> str:
         """
